Move this file into "constants" folder in my-map folder
