+++ pokemon-minisite/src/app/components/my-map/constants/my-map-style.ts
